# openapi/hashi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Hashi API",
    "description": "Hashi V2 admin and edge API. Frontend clients must consume this contract only.",
    "version": "2.0.0-alpha"
  },
  "paths": {
...
    "/api/vault/setup": {
      "post": {
        "tags": [
          "Vault"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vault/unlock": {
      "post": {
        "tags": [
          "Vault"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultUnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/vault/secrets": {
      "get": {
        "tags": [
          "Vault"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SecretDescriptorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vault"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vault/secrets/{id}/reveal": {
      "get": {
        "tags": [
          "Vault"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vault/verify-unlock": {
      "post": {
        "tags": [
          "Vault"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultUnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/dns/connections/{connectionId}/sync/plan": {
      "post": {
        "tags": [
          "DNS"
        ],
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DnsSyncPlanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dns/connections/{connectionId}/sync/apply": {
      "post": {
        "tags": [
          "DNS"
        ],
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsSyncApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sync/runs/{id}": {
      "get": {
        "tags": [
          "Sync"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sync/apply": {
      "post": {
        "tags": [
          "Sync"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DnsPlanChangeResponse": {
        "required": [
          "kind",
          "name",
          "type",
          "currentValue",
          "desiredValue",
          "ttl",
          "riskReason"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "currentValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "desiredValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "ttl": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "riskReason": {
            "type": "string"
          }
        }
      },
...
      "DnsSyncApplyRequest": {
        "required": [
          "planId",
          "connectionId",
          "confirmDestructive"
        ],
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "confirmDestructive": {
            "type": "boolean"
          }
        }
      },
      "DnsSyncPlanResponse": {
        "required": [
          "planId",
          "connectionId",
          "zoneName",
          "changes",
          "requiresConfirmation"
        ],
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneName": {
            "type": "string"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DnsPlanChangeResponse"
            }
          },
          "requiresConfirmation": {
            "type": "boolean"
          }
        }
      },
...
      "SecretDescriptorResponse": {
        "required": [
          "id",
          "purpose",
          "label",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "purpose": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SecretStoreRequest": {
        "required": [
          "label",
          "purpose",
          "plaintextBase64"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "plaintextBase64": {
            "type": "string"
          }
        }
      },
...
      "SyncApplyRequest": {
        "required": [
          "confirmDestructive"
        ],
        "type": "object",
        "properties": {
          "confirmDestructive": {
            "type": "boolean"
          }
        }
      },
...
      "VaultSetupRequest": {
        "required": [
          "recoveryKey",
          "prfWrapAttempted",
          "prfOutputBase64",
          "passkeyCredentialId"
        ],
        "type": "object",
        "properties": {
          "recoveryKey": {
            "type": "string"
          },
          "prfWrapAttempted": {
            "type": "boolean"
          },
          "prfOutputBase64": {
            "type": [
              "null",
              "string"
            ]
          },
          "passkeyCredentialId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "VaultUnlockRequest": {
        "required": [
          "recoveryKey"
        ],
        "type": "object",
        "properties": {
          "recoveryKey": {
            "type": "string"
          }
        }
      }
    }
  },
...
}